requirements.txt
PyGithub>=2.1.0
requests>=2.31.0
pandas>=2.0.0
jinja2>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
click>=8.1.0